pyproject.toml
[tool.poetry]
name = "nonebot-plugin-recall"
version = "0.1.0"
description = "Nonebot 2 的撤回插件"
authors = ["Yan <1964649083@qq.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
nonebot2 = "^2.4.0"
redis = "^5.2.0"


[tool.poetry.group.dev.dependencies]
nonebot-adapter-onebot = "^2.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
